admin product delete added

+++ src/Pages/Dashboard/ManageOrders.js
@@ -57,19 +57,22 @@ const ManageOrders = () => {
             })
     }
     const handleCancelOrder = _id => {
-        setAdminLoading(true);
-        fetch(`http://localhost:5000/orders/${_id}`, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    const remainingOrders = allOrders.filter(order => order._id !== _id);
-                    setAllOrders(remainingOrders);
-                    setAdminLoading(false);
-                    toast.success('Order Cancelled Successfully!');
-                }
+        const proceed = window.confirm('Are you sure to cancel?');
+        if (proceed) {
+            setAdminLoading(true);
+            fetch(`http://localhost:5000/orders/${_id}`, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount > 0) {
+                        const remainingOrders = allOrders.filter(order => order._id !== _id);
+                        setAllOrders(remainingOrders);
+                        setAdminLoading(false);
+                        toast.success('Order Cancelled Successfully!');
+                    }
+                })
+        }
     }
     return (
         <div>
